change: refactoring appointment module
