+++ backend/src/routes/users.routes.ts
@@ -53,6 +53,4 @@ router.delete(
   usersController.deleteUser
 );
 
-router.get("/appointments/:id",authenticate, usersController.getUserAppointments);
-
 export default router;
