+++ backend/src/service/users.service.ts
@@ -1,7 +1,7 @@
 import { prisma } from '../util/prisma.js';
 import { UserWithRoleAndPrivileges, UpdateUserRequest } from '../types/user.types';
 import { NotFoundError } from '../util/errors.util';
-import userModel from '../model/user.model';
+import userModel from '../model/appointment.model.js';
 
 /**
  * Get all users with their roles and privileges
@@ -122,11 +122,4 @@ export const getUserByUsername = async (username: string): Promise<UserWithRoleA
   });
 };
 
-/**
- * 
- * @param userId 
- * @returns List appointment
- */
-export const getAllAppointmentsByUserId = async (userId: string) => {
-  return await userModel.getAppointmentByUserId(userId);
-};
+
